feat(ui): Optimize input rendering with memoization
Refactors the `InputPrompt` component to improve typing performance. The text rendering logic has been extracted into a new, memoized `TextInputDisplay` component.

This change prevents the main `InputPrompt` component from re-rendering on every keystroke, addressing a performance bottleneck. By memoizing the display part, we ensure that only the text input itself is updated, while the more complex logic in the parent component is not re-evaluated unnecessarily.
